adding auth to the Tasks endPoints

+++ src/routers/task.js
@@ -17,20 +17,20 @@ router.post("/tasks", auth, async(req,res)=>{
     }
 })
 
-router.get("/tasks", async(req,res)=>{
+router.get("/tasks", auth, async(req,res)=>{
     try{
-        const tasks = await Task.find({})
-        res.status(200).send(tasks)
+        await req.user.populate("tasks")
+        res.status(200).send(req.user.tasks)
     }catch(err){
         res.status(500).send()
     }
 })
 
-router.get("/tasks/:id", async(req,res)=>{
+router.get("/tasks/:id", auth,async(req,res)=>{
     const _id=req.params.id
 
     try{
-        const task= await Task.findById(_id)
+        const task= await Task.findOne({_id, owner: req.user._id})
 
         if (!task) return res.status(404).send()
 
@@ -40,7 +40,7 @@ router.get("/tasks/:id", async(req,res)=>{
     }
 })
 
-router.patch("/tasks/:id", async (req,res)=>{
+router.patch("/tasks/:id", auth, async (req,res)=>{
 
     const updates= Object.keys(req.body)
     const allowedUpdates= ["description","completed"]
@@ -49,7 +49,7 @@ router.patch("/tasks/:id", async (req,res)=>{
     if (!isValidOperation) return res.status(400).send({error: "Updates are not allowed !"})
 
     try{     
-        const task = await Task.findById(req.params.id)
+        const task = await Task.findOne({_id: req.params.id, owner: req.user._id})
 
         if (!task) return res.status(404).send()
 
@@ -65,9 +65,9 @@ router.patch("/tasks/:id", async (req,res)=>{
 
 })
 
-router.delete("/tasks/:id", async(req,res)=>{
+router.delete("/tasks/:id", auth, async(req,res)=>{
     try{
-        const task= await Task.findOneAndDelete(req.params.id)
+        const task= await Task.findOneAndDelete({_id: req.params.id, owner: req.user._id})
 
         if(!task) return res.status(404).send()
